Moves session factory into models module
